fix: repair CI/build and add generic shard keys (client sharding gap #1)

+++ sharding-jdbc/src/main/java/org/springframework/boot/starter/sharding/jdbc/RoutingDataSource.java
@@ -52,7 +52,9 @@ public class RoutingDataSource extends AbstractDataSource {
             throw new MissingShardKeyException();
         }
 
-        Shard shard = shardRouter.resolve(shardKey);
+        // Call the primitive resolve(long) explicitly: passing the boxed Long would
+        // bind to the default resolve(Object) overload instead.
+        Shard shard = shardRouter.resolve(shardKey.longValue());
         return shard.dataSource();
     }
